feat(misc): add --integrated option to nx init command
