+++ packages/nx/src/command-line/init.ts
@@ -13,7 +13,11 @@ import { runNxSync } from '../utils/child-process';
 import { directoryExists, readJsonFile } from '../utils/fileutils';
 import { PackageJson } from '../utils/package-json';
 
-export async function initHandler() {
+export interface InitArgs {
+  integrated: boolean;
+}
+
+export async function initHandler(options: InitArgs) {
   const args = process.argv.slice(2).join(' ');
   const flags = parser(args, {
     boolean: ['useDotNxInstallation'],
@@ -36,9 +40,9 @@ export async function initHandler() {
   } else if (existsSync('package.json')) {
     const packageJson: PackageJson = readJsonFile('package.json');
     if (existsSync('angular.json')) {
-      await addNxToAngularCliRepo();
+      await addNxToAngularCliRepo(options.integrated);
     } else if (isCRA(packageJson)) {
-      await addNxToCraRepo();
+      await addNxToCraRepo(options.integrated);
     } else if (isNestCLI(packageJson)) {
       await addNxToNest(packageJson);
     } else if (isMonorepo(packageJson)) {
