+++ packages/nx/src/nx-init/angular/index.ts
@@ -1,26 +1,19 @@
 import { prompt } from 'enquirer';
-import { unlinkSync } from 'fs';
-import { dirname, join, relative, resolve } from 'path';
-import { toNewFormat } from '../../adapter/angular-json';
-import { NxJsonConfiguration } from '../../config/nx-json';
-import {
-  ProjectConfiguration,
-  TargetConfiguration,
-} from '../../config/workspace-json-project-json';
-import { fileExists, readJsonFile, writeJsonFile } from '../../utils/fileutils';
+import { join } from 'path';
+import { readJsonFile, writeJsonFile } from '../../utils/fileutils';
 import { sortObjectByKeys } from '../../utils/object-sort';
 import { output } from '../../utils/output';
-import { PackageJson } from '../../utils/package-json';
-import { normalizePath } from '../../utils/path';
+import type { PackageJson } from '../../utils/package-json';
 import {
   addDepsToPackageJson,
-  addVsCodeRecommendedExtensions,
   askAboutNxCloud,
-  createNxJsonFile,
   initCloud,
   runInstall,
 } from '../utils';
+import { setupIntegratedWorkspace } from './integrated-workspace';
 import { getLegacyMigrationFunctionIfApplicable } from './legacy-angular-versions';
+import { setupStandaloneWorkspace } from './standalone-workspace';
+import type { AngularJsonConfig } from './types';
 import yargsParser = require('yargs-parser');
 
 const defaultCacheableOperations: string[] = [
@@ -40,12 +33,13 @@ const parsedArgs = yargsParser(process.argv, {
 let repoRoot: string;
 let workspaceTargets: string[];
 
-export async function addNxToAngularCliRepo() {
+export async function addNxToAngularCliRepo(integrated: boolean) {
   repoRoot = process.cwd();
 
   output.log({ title: '🧐 Checking versions compatibility' });
   const legacyMigrationFn = await getLegacyMigrationFunctionIfApplicable(
     repoRoot,
+    integrated,
     parsedArgs.yes !== true
   );
   if (legacyMigrationFn) {
@@ -60,14 +54,16 @@ export async function addNxToAngularCliRepo() {
   });
 
   output.log({ title: '🐳 Nx initialization' });
-  const cacheableOperations = await collectCacheableOperations();
+  const cacheableOperations = !integrated
+    ? await collectCacheableOperations()
+    : [];
   const useNxCloud = parsedArgs.yes !== true ? await askAboutNxCloud() : false;
 
   output.log({ title: '📦 Installing dependencies' });
   installDependencies(useNxCloud);
 
   output.log({ title: '📝 Setting up workspace' });
-  await setupWorkspace(cacheableOperations);
+  await setupWorkspace(cacheableOperations, integrated);
 
   if (useNxCloud) {
     output.log({ title: '🛠️ Setting up Nx Cloud' });
@@ -155,151 +151,19 @@ function addPluginDependencies(): void {
   writeJsonFile(packageJsonPath, packageJson);
 }
 
-type AngularJsonConfigTargetConfiguration = Exclude<
-  TargetConfiguration,
-  'command' | 'executor' | 'outputs' | 'dependsOn' | 'inputs'
-> & {
-  builder: string;
-};
-type AngularJsonProjectConfiguration = {
-  root: string;
-  sourceRoot: string;
-  architect?: Record<string, AngularJsonConfigTargetConfiguration>;
-};
-interface AngularJsonConfig {
-  projects: Record<string, AngularJsonProjectConfiguration>;
-  defaultProject?: string;
-}
-
-async function setupWorkspace(cacheableOperations: string[]): Promise<void> {
-  const angularJsonPath = join(repoRoot, 'angular.json');
-  const angularJson = readJsonFile<AngularJsonConfig>(angularJsonPath);
-  const workspaceCapabilities = getWorkspaceCapabilities(angularJson.projects);
-  createNxJson(angularJson, cacheableOperations, workspaceCapabilities);
-  addVsCodeRecommendedExtensions(
-    repoRoot,
-    [
-      'nrwl.angular-console',
-      'angular.ng-template',
-      workspaceCapabilities.eslintProjectConfigFile
-        ? 'dbaeumer.vscode-eslint'
-        : undefined,
-    ].filter(Boolean)
-  );
-  replaceNgWithNxInPackageJsonScripts();
-
-  // convert workspace config format to standalone project configs
-  // update its targets outputs and delete angular.json
-  const projects = toNewFormat(angularJson).projects;
-  for (const [projectName, project] of Object.entries(projects)) {
-    updateProjectOutputs(project);
-    writeJsonFile(join(project.root, 'project.json'), {
-      $schema: normalizePath(
-        relative(
-          join(repoRoot, project.root),
-          join(repoRoot, 'node_modules/nx/schemas/project-schema.json')
-        )
-      ),
-      name: projectName,
-      ...project,
-      root: undefined,
-    });
-  }
-  unlinkSync(angularJsonPath);
-}
-
-type WorkspaceCapabilities = {
-  eslintProjectConfigFile: boolean;
-  test: boolean;
-  karmaProjectConfigFile: boolean;
-};
-
-function createNxJson(
-  angularJson: AngularJsonConfig,
+async function setupWorkspace(
   cacheableOperations: string[],
-  {
-    eslintProjectConfigFile,
-    test,
-    karmaProjectConfigFile,
-  }: WorkspaceCapabilities
-): void {
-  createNxJsonFile(repoRoot, [], cacheableOperations, {});
-
-  const nxJson = readJsonFile<NxJsonConfiguration>(join(repoRoot, 'nx.json'));
-  nxJson.namedInputs = {
-    sharedGlobals: [],
-    default: ['{projectRoot}/**/*', 'sharedGlobals'],
-    production: [
-      'default',
-      ...(test
-        ? [
-            '!{projectRoot}/tsconfig.spec.json',
-            '!{projectRoot}/**/*.spec.[jt]s',
-            karmaProjectConfigFile ? '!{projectRoot}/karma.conf.js' : undefined,
-          ].filter(Boolean)
-        : []),
-      eslintProjectConfigFile ? '!{projectRoot}/.eslintrc.json' : undefined,
-    ].filter(Boolean),
-  };
-  nxJson.targetDefaults = {};
-  if (workspaceTargets.includes('build')) {
-    nxJson.targetDefaults.build = {
-      dependsOn: ['^build'],
-      inputs: ['production', '^production'],
-    };
+  isIntegratedMigration: boolean
+): Promise<void> {
+  if (isIntegratedMigration) {
+    setupIntegratedWorkspace();
+  } else {
+    await setupStandaloneWorkspace(
+      repoRoot,
+      cacheableOperations,
+      workspaceTargets
+    );
   }
-  if (workspaceTargets.includes('server')) {
-    nxJson.targetDefaults.server = { inputs: ['production', '^production'] };
-  }
-  if (workspaceTargets.includes('test')) {
-    const inputs = ['default', '^production'];
-    if (fileExists(join(repoRoot, 'karma.conf.js'))) {
-      inputs.push('{workspaceRoot}/karma.conf.js');
-    }
-    nxJson.targetDefaults.test = { inputs };
-  }
-  if (workspaceTargets.includes('lint')) {
-    const inputs = ['default'];
-    if (fileExists(join(repoRoot, '.eslintrc.json'))) {
-      inputs.push('{workspaceRoot}/.eslintrc.json');
-    }
-    nxJson.targetDefaults.lint = { inputs };
-  }
-  if (workspaceTargets.includes('e2e')) {
-    nxJson.targetDefaults.e2e = { inputs: ['default', '^production'] };
-  }
-  // Angular 14 workspaces support defaultProject, keep it until we drop support
-  nxJson.defaultProject = angularJson.defaultProject;
-  writeJsonFile(join(repoRoot, 'nx.json'), nxJson);
-}
-
-function updateProjectOutputs(project: ProjectConfiguration): void {
-  Object.values(project.targets ?? {}).forEach((target) => {
-    if (
-      [
-        '@angular-devkit/build-angular:browser',
-        '@angular-builders/custom-webpack:browser',
-        'ngx-build-plus:browser',
-        '@angular-devkit/build-angular:server',
-        '@angular-builders/custom-webpack:server',
-        'ngx-build-plus:server',
-      ].includes(target.executor)
-    ) {
-      target.outputs = ['{options.outputPath}'];
-    } else if (target.executor === '@angular-eslint/builder:lint') {
-      target.outputs = ['{options.outputFile}'];
-    } else if (target.executor === '@angular-devkit/build-angular:ng-packagr') {
-      try {
-        const ngPackageJsonPath = join(repoRoot, target.options.project);
-        const ngPackageJson = readJsonFile(ngPackageJsonPath);
-        const outputPath = relative(
-          repoRoot,
-          resolve(dirname(ngPackageJsonPath), ngPackageJson.dest)
-        );
-        target.outputs = [`{workspaceRoot}/${normalizePath(outputPath)}`];
-      } catch {}
-    }
-  });
 }
 
 function getWorkspaceTargets(): string[] {
@@ -314,75 +178,4 @@ function getWorkspaceTargets(): string[] {
   }
 
   return Array.from(targets);
-}
-
-function getWorkspaceCapabilities(
-  projects: Record<string, AngularJsonProjectConfiguration>
-): WorkspaceCapabilities {
-  const capabilities = {
-    eslintProjectConfigFile: false,
-    test: false,
-    karmaProjectConfigFile: false,
-  };
-
-  for (const project of Object.values(projects)) {
-    if (
-      !capabilities.eslintProjectConfigFile &&
-      projectHasEslintConfig(project)
-    ) {
-      capabilities.eslintProjectConfigFile = true;
-    }
-    if (!capabilities.test && projectUsesKarmaBuilder(project)) {
-      capabilities.test = true;
-    }
-    if (
-      !capabilities.karmaProjectConfigFile &&
-      projectHasKarmaConfig(project)
-    ) {
-      capabilities.karmaProjectConfigFile = true;
-    }
-
-    if (
-      capabilities.eslintProjectConfigFile &&
-      capabilities.test &&
-      capabilities.karmaProjectConfigFile
-    ) {
-      return capabilities;
-    }
-  }
-
-  return capabilities;
-}
-
-function projectUsesKarmaBuilder(
-  project: AngularJsonProjectConfiguration
-): boolean {
-  return Object.values(project.architect ?? {}).some(
-    (target) => target.builder === '@angular-devkit/build-angular:karma'
-  );
-}
-
-function projectHasKarmaConfig(
-  project: AngularJsonProjectConfiguration
-): boolean {
-  return fileExists(join(project.root, 'karma.conf.js'));
-}
-
-function projectHasEslintConfig(
-  project: AngularJsonProjectConfiguration
-): boolean {
-  return fileExists(join(project.root, '.eslintrc.json'));
-}
-
-function replaceNgWithNxInPackageJsonScripts(): void {
-  const packageJsonPath = join(repoRoot, 'package.json');
-  const packageJson = readJsonFile<PackageJson>(packageJsonPath);
-  packageJson.scripts ??= {};
-  Object.keys(packageJson.scripts).forEach((script) => {
-    packageJson.scripts[script] = packageJson.scripts[script]
-      .replace(/^nx$/, 'nx')
-      .replace(/^ng /, 'nx ')
-      .replace(/ ng /g, ' nx ');
-  });
-  writeJsonFile(packageJsonPath, packageJson);
 }
