+++ packages/nx/src/nx-init/angular/integrated-workspace.ts
@@ -1,0 +1,7 @@
+import { execSync } from 'child_process';
+import { getPackageManagerCommand } from '../../utils/package-manager';
+
+export function setupIntegratedWorkspace(): void {
+  const pmc = getPackageManagerCommand();
+  execSync(`${pmc.exec} ng g @nrwl/angular:ng-add`, { stdio: [0, 1, 2] });
+}
